Re-implement websocket support
Re-implement websocket support to handle JSON RPC requests.
This implies the usage of `asyncio` and moves most of the client
function to be `async`.
To do this the old `jsonrpcclient` and `websocket-client`
dependencies were dropped in favor of `fastapi_websocket_rpc` which
required little modifications to make it work with the Nimiq server,
namely (see `websocket_rpc.py`):
- Custom JSON serializers.
- Description of the methods that the server can request which for
  now only includes `headSubscribe`.

Also:
- Move the examples from `example` to `examples`.
- Rename `example.py` to `basic_example.py` and adapted example to
  the new way of using the client.
- Add an example that subscribes to blocks being produced:
  `subscription_example.py`